Initial Commit
Right now, rcfile is in a working state. Documentation is non-existent other
than the incomplete OptionParser. There are various environment variables that
can be set to alter the behavior of the program. None of these are documented
currently.

Goals:

-Add automatic rcfile updating through pexpect as an option in the OptionParser.
The pexpect code works and is included, it just hasn't been added to the command
line interface yet.

-Allow updating through webtiles websocket connections for cao and cszo. Players
on cdo will have to use the DGL interface for now. rwbarton suggested using a
library called wssh for interfacing with the webtiles interface over websockets:
https://github.com/progrium/wssh

-Add 'output to file' and 'open in browser' options.

-Write documentation.